MyRequestLine클래스 작성
- requestLine에 대한 책임을 가지는 MyRequestLine 클래스를 분리했다.
